+++ release-notes.md
@@ -5,6 +5,7 @@ RELEASE NOTES
 - add support for css selector when setExcludeClasses
 - update webdriver version 2.49.0
 - add log at sendKeys method
+- add possibility to overwritten value from webLocator.properties in browser.properties
 
 **Release Notes for Testy 2.0.0**
 - Change config default value for weblocator.log.suggestions=false